requirements.txt
pillow
pytesseract
tesserocr
weighted-levenshtein
numpy
scikit-image
